Fixing/improving build script
I had misread a build failure and had assumed that the static directory
had been removed. It was actually the public directory as part of the
docusaurus migration. As a result, I had updated this build script to
instead build directly into the build/ folder instead of static.
However, the lingering copy of the code in static was _overwriting_ the
latest build. So, I've re-updated this script to copy back into static
which is loading the latest copy of the game now. Moreover, I updated
the script to DRY up some things and make it clearer what the script is
doing.